isolate queue handling to queue saga
